refactor(LeadFinalApprovalExecutor): extract domain modules for maintainability
Refactored LeadFinalApprovalExecutor.js (879 LOC) into focused domain modules:

- gates.js (461 LOC): Gate validator definitions (PLAN_TO_LEAD_HANDOFF_EXISTS, USER_STORIES_COMPLETE, RETROSPECTIVE_EXISTS, PR_MERGE_VERIFICATION)
- helpers.js (251 LOC): Helper methods (checkAndCompleteParentSD, resolveLearningItems, releaseSessionClaim)
- remediations.js (64 LOC): Remediation messages for failed gates
- index.js (186 LOC): Main executor orchestrator with re-exports

Original file reduced to thin wrapper (34 LOC) maintaining backward compatibility.
Part of SD-LEO-REFACTOR-LARGE-FILES-003 orchestrator.
